Add additional significance information to README
I was on a run and thought about what beliefs are actually interesting
to write about and realized they don't have to be anywhere near
profound.

I was also thinking about identification schemes which led me to think
of how I could graph the relationships between beliefs and component
beliefs.