Add belief about benefits of social media
This one is a bit soft. I was trying to probe in to an easy belief. It
feels difficult to justify a belief that is strongly dependent on my
understanding of a single word "empathy". I believe empathy is good.
Why? Is there research I can cite that proves that people are better to
each other if they're more empathetic?

I'd also like to explore why I think increased narcissism is bad.